requirements.txt
Django==1.11
psycopg2
djangorestframework
pylint
autopep8